fixed bug

+++ OrderManagementSystem_WebAPI/Models/Order.cs
@@ -21,7 +21,7 @@ namespace OrderManagementSystem_WebAPI.Models
         [Required(ErrorMessage = "Customer Id Required")]
         public int CustomerId { get; set; }
 
-        public double? TotalAmount { get; set; }
+        public decimal? TotalAmount { get; set; }
 
         public Customer Customers { get; set; }
     }
